pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = "cute project"
authors = ["kpomak <romka_ne@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.16"
scrapy = "^2.7.1"
pymongo = "^4.3.3"
black = "^22.12.0"
python-dotenv = "^0.21.0"
isort = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
